+++ build.sh
@@ -33,8 +33,8 @@ cat > "$CONTENTS/Info.plist" <<'PLIST'
     <key>CFBundleDisplayName</key>      <string>Cookie Monster</string>
     <key>CFBundleIdentifier</key>       <string>com.pflugpeil.cookiemonster</string>
     <key>CFBundleExecutable</key>       <string>CookieMonster</string>
-    <key>CFBundleVersion</key>          <string>0.3.0</string>
-    <key>CFBundleShortVersionString</key><string>0.3.0</string>
+    <key>CFBundleVersion</key>          <string>0.4.0</string>
+    <key>CFBundleShortVersionString</key><string>0.4.0</string>
     <key>CFBundlePackageType</key>      <string>APPL</string>
     <key>CFBundleIconFile</key>         <string>AppIcon</string>
     <key>LSMinimumSystemVersion</key>   <string>13.0</string>